refactor: move mock implementations to class
