+++ src/user/__mocks__/user.ts
@@ -1,0 +1,40 @@
+import { faker } from '@faker-js/faker';
+
+class MockUser {
+  email: string;
+  password: string;
+  role?: string;
+}
+
+export class mockPrismaService {
+  public user = {
+    create: function (request: { data: MockUser }) {
+      return Promise.resolve({
+        id: faker.number.int(),
+        email: request.data.email,
+        password: request.data.password,
+        role: 'USER',
+      });
+    },
+    update: function (request: {
+      data: {
+        email: string;
+        password: string;
+      };
+      where: {
+        id: number;
+      };
+    }) {
+      return Promise.resolve({
+        id: request.where.id,
+        email: request.data.email,
+        password: request.data.password,
+        role: 'USER',
+      });
+    },
+
+    delete: function (_request: { id: number }) {
+      return Promise.resolve();
+    },
+  };
+}
